// carbookingfrontend/carbookingfrontend/src/app/models/customermodel.ts
export class customerModel {
    "Firstname": any;
    "Lastname": any;
    "Emailid": any;
    "pwd": any;
    "DOB": any;
    "Gender": any;
    "Address": any;
    "City": any;
    "State": any;
    "Pincode": any;
    "Country": any;
    "Contactnumber": any;
    "Whatsupnumber": any;
    "Aadharnumber": any;
    "IsCustomer": any;
    }

    export class customerLoginModel {
        "Emailid": any;
        "pwd": any;
}